+++ GeosCore/hco_utilities_gc_mod.F90
@@ -2573,6 +2573,13 @@ CONTAINS
       ! Free pointer
       SpcInfo => NULL()
 
+      ! Each time the boundary conditions are read, they have no longer been perturbed
+      ! This value is sometimes set to True in set_boundary_conditions_mod.F90
+      IF ( ( Input_Opt%ITS_A_CH4_SIM .OR. Input_Opt%ITS_A_CARBON_SIM ) .AND. & 
+           Input_Opt%PerturbCH4BoundaryConditions ) THEN
+        State_Chm%IsCH4BCPerturbed = .FALSE.
+      ENDIF
+
    ENDDO
 
    ! Reset FIRST flag